+++ webpack/webpack.common.js
@@ -42,10 +42,11 @@ module.exports = {
       {
         test: /\.(ico|jpg|jpeg|png|gif|eot|otf|webp|svg|ttf|woff|woff2)(\?.*)?$/,
         use: {
-          loader: 'file-loader',
-          options: {
-            name: '[path][name].[ext]',
-          },
+          loader: 'file-loader?name=img/[path][name].[ext]&context=./src/img',
+          // loader: 'file-loader',
+          // options: {
+          //   name: '[path][name].[ext]',
+          // },
         },
       },
     ],